Use nulled variable in Request get_dict() method instead of self._nullable
Request.get_dict currently uses the class variable of self._nullable after creating a local variable nulled from the Request.get_nulled method. This passes all nullable fields to the get_dict method and leaves them missing from the returned dict.

Currently, the default nullable parameter in util.todict is set to an empty list which is a mutable object.

Changed to set it to None, then check if it's equal to None before creating it. This doesn't currently cause any issues.

See:
https://docs.python.org/3/reference/compound_stmts.html#function-definitions